+++ src/sub-clients/accounts/client/Account.test.pact.ts
@@ -3,10 +3,13 @@ import { provider } from '../../../pact-support/pactSetup';
 import { withClients } from '../../../pact-support/pactTestWrapper';
 import { FakeBoclipsClient, isATestClient } from '../../../test-support';
 import { AccountFactory } from '../../../test-support/AccountFactory';
-import { getAccountsByCountryCode } from '../pact/AccountInteractions';
+import { UpdateAccountRequest } from '../model/UpdateAccountRequest';
+import {
+  getAccountsByCountryCode,
+  updateAccount,
+} from '../pact/AccountInteractions';
 
 const USA_ACCOUNT_ID_FROM_STAGING = '5de5369d1617b3000142c08f';
-const GBR_ACCOUNT_ID_FROM_STAGING = '5d7b606e861a170001e11dc9';
 
 describe('Account', () => {
   withClients(
@@ -20,11 +23,6 @@ describe('Account', () => {
           client.accountsClient.insertAccountFixture(
             AccountFactory.sample({
               id: USA_ACCOUNT_ID_FROM_STAGING,
-            }),
-          );
-          client.accountsClient.insertAccountFixture(
-            AccountFactory.sample({
-              id: GBR_ACCOUNT_ID_FROM_STAGING,
             }),
           );
         }
@@ -49,21 +47,46 @@ describe('Account', () => {
       });
 
       it('can fetch all accounts', async () => {
-        if (isATestClient(await client)) {
-          const accountsPage = await client.accountsClient.getAccounts();
+        await provider.addInteraction(
+          getAccountsByCountryCode(USA_ACCOUNT_ID_FROM_STAGING, 'USA'),
+        );
+        const accountsPage = await client.accountsClient.getAccounts({
+          countryCode: 'USA',
+          page: 0,
+          size: 30,
+        });
 
-          expect(accountsPage.page.length).toBeGreaterThanOrEqual(2);
-          expect(
-            accountsPage.page.some(
-              account => account.id === USA_ACCOUNT_ID_FROM_STAGING,
-            ),
-          ).toBeTruthy();
-          expect(
-            accountsPage.page.some(
-              account => account.id === GBR_ACCOUNT_ID_FROM_STAGING,
-            ),
-          ).toBeTruthy();
-        }
+        expect(accountsPage.page.length).toBeGreaterThanOrEqual(1);
+        expect(
+          accountsPage.page.some(
+            account => account.id === USA_ACCOUNT_ID_FROM_STAGING,
+          ),
+        ).toBeTruthy();
+      });
+
+      it(`can edit an account's access expiry date`, async () => {
+        const updatedTime = '2020-12-21T00:00:00.000';
+        const updateRequest: UpdateAccountRequest = {
+          accessExpiresOn: new Date(updatedTime),
+        };
+        await provider.addInteraction(
+          getAccountsByCountryCode(USA_ACCOUNT_ID_FROM_STAGING, 'USA'),
+        );
+        await provider.addInteraction(
+          updateAccount(USA_ACCOUNT_ID_FROM_STAGING, updateRequest),
+        );
+        const accountPage = await client.accountsClient.getAccounts({
+          countryCode: 'USA',
+          page: 0,
+          size: 30,
+        });
+
+        const updatedAccount = await client.accountsClient.updateAccount(
+          accountPage.page[0],
+          updateRequest,
+        );
+
+        expect(updatedAccount.accessExpiresOn).toEqual(new Date(updatedTime));
       });
     },
   );