+++ src/sub-clients/accounts/client/AccountsClient.ts
@@ -1,5 +1,6 @@
 import Pageable from '../../common/model/Pageable';
 import { Account } from '../model/Account';
+import { UpdateAccountRequest } from '../model/UpdateAccountRequest';
 
 export interface AccountsFilter {
   countryCode?: string;
@@ -11,4 +12,9 @@ export interface AccountsClient {
   getAccounts(
     accountsFilter?: AccountsFilter,
   ): Promise<Pageable<Account> | null>;
+
+  updateAccount(
+    account: Account,
+    updateAccountRequest: UpdateAccountRequest,
+  ): Promise<Account>;
 }