+++ src/sub-clients/accounts/model/UpdateAccountRequest.ts
@@ -1,0 +1,3 @@
+export class UpdateAccountRequest {
+  accessExpiresOn: Date;
+}