+++ src/sub-clients/accounts/pact/AccountInteractions.ts
@@ -1,5 +1,7 @@
 import { InteractionObject } from '@pact-foundation/pact';
 import { eachLike, like } from '@pact-foundation/pact/dsl/matchers';
+import { provider } from '../../../pact-support/pactSetup';
+import { UpdateAccountRequest } from '../model/UpdateAccountRequest';
 
 export const getAccountsByCountryCode = (
   id: string,
@@ -49,10 +51,50 @@ export const getAccountsByCountryCode = (
           }),
           _links: like({
             edit: {
-              href:
-                'https://api.boclips.com/v1/accounts/5daa05af7e5bb50001ead980',
+              href: `${provider.mockService.baseUrl}/v1/accounts/${id}`,
             },
           }),
+        }),
+      },
+    },
+  },
+});
+
+export const updateAccount = (
+  id: string,
+  updateAccountRequest: UpdateAccountRequest,
+): InteractionObject => ({
+  state: undefined,
+  uponReceiving: 'PATCH Account to change accessExpiresOn',
+  withRequest: {
+    method: 'PATCH',
+    path: `/v1/accounts/${id}`,
+    headers: {
+      'Content-Type': 'application/json;charset=utf-8',
+    },
+    body: updateAccountRequest,
+  },
+  willRespondWith: {
+    status: 200,
+    headers: {
+      'Content-Type': 'application/hal+json;charset=UTF-8',
+    },
+    body: {
+      id: like(id),
+      accessExpiresOn: like(updateAccountRequest.accessExpiresOn.toISOString()),
+      contractIds: like([]),
+      organisation: like({
+        name: like('1st Football High School'),
+        type: like('SCHOOL'),
+        country: like({
+          id: `USA`,
+          name: 'United States',
+          states: null,
+        }),
+      }),
+      _links: {
+        edit: like({
+          href: `${provider.mockService.baseUrl}/v1/accounts/${id}`,
         }),
       },
     },